refactor: 외부 API 연동 중 환율 처리 예외 패키지 이동
