+++ src/main/java/com/wanderpass/exchange_rate_service/exchange/exception/exchange/ExchangeRateApiException.java
@@ -1,7 +1,7 @@
 package com.wanderpass.exchange_rate_service.exchange.exception.exchange;
 
-import com.wanderpass.exchange_rate_service.exchange.exception.type.IntegrationException;
-import com.wanderpass.exchange_rate_service.exchange.exception.type.StatusCode;
+import com.wanderpass.exchange_rate_service.global.exception.type.IntegrationException;
+import com.wanderpass.exchange_rate_service.global.exception.type.StatusCode;
 
 public class ExchangeRateApiException extends IntegrationException {
     public ExchangeRateApiException(StatusCode statusCode, String message) {
